update: 設定コマンドのmongoDB対応 & リファクタリング
