--- commands/setting/buttons/report/reportRole.js
@@ -1,28 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-const discord = require('discord.js');
-
-/** @type {import('@djs-tools/interactions').ButtonRegister} */
-const ping_command = {
-    data: {
-        customId: 'setting-reportRole',
-        type: 'BUTTON',
-    },
-    exec: async (interaction) => {
-        const modal = new discord.ModalBuilder()
-            .setCustomId('setting-Role')
-            .setTitle('ロールメンション')
-            .addComponents(
-                new discord.ActionRowBuilder().addComponents(
-                    new discord.TextInputBuilder()
-                        .setCustomId('reportRole')
-                        .setLabel('ロールの名前')
-                        .setMaxLength(100)
-                        .setStyle(discord.TextInputStyle.Short)
-                        .setRequired(true),
-                ),
-            );
-
-        interaction.showModal(modal);
-    },
-};
-module.exports = [ ping_command ];
