--- commands/setting/buttons/report/reportRoleMention.js
@@ -1,36 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-const discord = require('discord.js');
-const { settingSwitcher } = require('../../../../modules/switcher');
-
-/** @type {import('@djs-tools/interactions').ButtonRegister} */
-const ping_command = {
-    data: {
-        customId: 'setting-reportRoleMention',
-        type: 'BUTTON',
-    },
-    exec: async (interaction) => {
-        const Model = await require('../../../../models/basic')(interaction.sequelize).findOne({ where: { serverId: interaction.guildId } });
-        const { reportRole, reportRoleMention } = Model.get();
-
-        const embed = interaction.message.embeds[0];
-        const button = interaction.message.components[1];
-
-        let err = false;
-        Model.update({ reportRoleMention: reportRoleMention ? false : true }).catch(() => err = true);
-
-        if (err) {
-            const error = new discord.EmbedBuilder()
-                .setDescription('❌ 設定を正しく保存できませんでした。時間を置いて再試行してください。')
-                .setColor('Red');
-            return interaction.update({ embeds: [embed, error] });
-        }
-
-        embed.fields[1].value = settingSwitcher('STATUS_ROLE', !reportRoleMention, reportRole);
-        button.components[1] = discord.ButtonBuilder.from(button.components[1])
-            .setLabel(settingSwitcher('BUTTON_LABEL', !reportRoleMention))
-            .setStyle(settingSwitcher('BUTTON_STYLE', !reportRoleMention));
-
-        interaction.update({ embeds: [embed], components: [interaction.message.components[0], button] });
-    },
-};
-module.exports = [ ping_command ];
