+++ modules/settingStatusSwitcher.js
